Add procedural scene generator and refactor logging includes
Introduces SceneGenerator for procedural voxel scene generation with Cornell Box, cave, and urban grid support. Refactors logger includes to use local paths, adds logger initialization to relevant classes, and updates SpirvReflector to be non-static. Minor type and usage fixes in ShaderManagement modules.

+++ VIXEN/ShaderManagement/src/AsyncShaderBundleBuilder.cpp
@@ -1,4 +1,5 @@
 #include "ShaderManagement/AsyncShaderBundleBuilder.h"
+#include "Logger.h"
 #include <algorithm>
 
 namespace ShaderManagement {
@@ -262,7 +263,7 @@ void AsyncShaderBundleBuilder::ExecuteBuild(
 
     std::string uuid = builder.GetUuid();
     std::string programName = builder.GetProgramName();
-    uint32_t stageCount = builder.GetStageCount();
+    uint32_t stageCount = static_cast<uint32_t>(builder.GetStageCount());
 
     // Check if cancelled
     {
